Rename interceptors to middleware
